+++ instater/context.py
@@ -1,19 +1,24 @@
+import os.path
+import time
 import typing
 from collections import Counter
 from pathlib import Path
-from typing import Iterable, List
+from typing import Iterable
 
 from jinja2 import Environment, FileSystemLoader
 from rich.console import Console
 
-import instater
-
 from . import util
+
+
+def _filename(path: str) -> str:
+    return os.path.basename(path).rsplit(".", 1)[0]
 
 
 def _jinja_environment(root_directory: Path) -> Environment:
     env = Environment(loader=FileSystemLoader(root_directory))
     env.filters["password_hash"] = util.password_hash
+    env.filters["filename"] = _filename
     return env
 
 
@@ -23,12 +28,14 @@ class Context:
         self.tags = set(tags)
         self.dry_run = dry_run
 
-        extra_vars["instater_dir"] = str(root_directory)
+        extra_vars["instater_dir"] = str(root_directory.absolute())
         self.variables = extra_vars
 
         self.jinja_env = _jinja_environment(root_directory)
-        self.tasks: List[instater.tasks.Task] = []
+        self.tasks: list = []
         self.statuses: typing.Counter[str] = Counter()
+
+        self.start = time.time()
 
         self.console = Console()
         self.print = self.console.print
@@ -56,11 +63,18 @@ class Context:
             vars = {**vars, **extra_vars}
         return self.jinja_env.get_template(template_path).render(vars)
 
+    def duration(self, start: float = None):
+        if start is None:
+            start = self.start
+
+        duration = round(time.time() - start, 3)
+        return f"[white]({duration}s)[/white]"
+
     def print_summary(self):
         skipped = self.statuses["skipped"]
         changed = self.statuses["changed"]
 
-        self.print("Summary:", style="bold")
+        self.print(f"Summary {self.duration()}:", style="bold")
         self.print(f"  skipped: {skipped}", style="blue")
         # if this used style="yellow", the integer count would be turned blue by rich
         self.print(f"  [yellow]changed: {changed}[/yellow]")